requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
matplotlib>=3.6.0
tqdm>=4.65.0
nibabel>=5.0.0  # For medical image I/O
SimpleITK>=2.2.0  # For medical image processing
scikit-image>=0.20.0
tensorboard>=2.12.0
einops>=0.6.0
pytest>=7.0.0
pytest-cov>=4.0.0 